refactor: move ocaasionNotification file to fds widget folder (#185)
