+++ src/fsd_widgets/notification/index.ts
@@ -4,5 +4,5 @@ export { NotificationSettingWidget } from './ui/NotificationSettingWidget';
 
 // ui
 export { NotificationWidget } from './ui/NotificationWidget';
-
+export { OccasionNotification } from './ui/OccasionNotification';
 export { FCMTokenChecker } from './ui/FCMTokenChecker';
