+++ src/fsd_widgets/notification/ui/index.ts
@@ -1,0 +1,1 @@
+export { OccasionNotification } from './OccasionNotification';
